backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Presidio
presidio-analyzer==2.2.333
presidio-anonymizer==2.2.333
spacy==3.7.2

# Encryption
cryptography==41.0.7

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2
requests==2.31.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1

